added the ability to switch platforms which is still in beta
